// src/app/core/models/dataLocalResp.interfaces.ts
// Generated by https://quicktype.io

export interface Data {
    mensaje_return: MensajeReturn;
    token:          string;
    result:         Result;
    menu_user:      Array<Array<MenuUserClass[] | string>>;
    data_company:   DataCompany;
}

export interface DataCompany {
    id_user_company:        number;
    id_user:                number;
    id_language:            number;
    id_user_type:           number;
    user_type:              string;
    status_user:            number;
    accept_term_condition:  number;
    id_registration_method: number;
    first_name:             string;
    surname:                string;
    alias_username:         string;
    email:                  string;
    email_validate_code:    string;
    validate_mail:          number;
    email_valid_until:      string;
    email_validated_on:     string;
    token:                  string;
    password:               string;
    change_password:        number;
    connection_status:      number;
    ip_last_connection:     string;
    conexion_date:          string;
    registration_date:      string;
    id_company:             number;
    id_courier:             number;
    address:                string;
    company_name:           string;
    img_logo:               string;
    id_company_type:        number;
    company_type:           string;
}

export interface MensajeReturn {
    ERROR:          boolean;
    ERROR_MENSAGGE: string;
    CODE:           number;
}

export interface MenuUserClass {
    menu_name:      string;
    icon_menu:      string;
    submenu_name:   string;
    route_sub_menu: string;
}

export interface Result {
    id_user:          number;
    first_name:       string;
    surname:          string;
    alias_username:   string;
    email:            string;
    validate_mail:    number;
    conexion_date:    string;
    id_language:      number;
    acronym_language: string;
    language_name:    string;
    icon:             null;
    id_user_type:     number;
    user_type:        string;
}